+++ assets/javascripts/kitten/components/menus/header-menu/test.js
@@ -87,4 +87,26 @@ describe('<HeaderMenu />', () => {
       expect(component).toMatchSnapshot()
     })
   })
+
+  describe('with button items', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(
+          <HeaderMenu borderSide="right">
+            <HeaderMenu.Item href="#">Item 1</HeaderMenu.Item>
+            <HeaderMenu.Item tag="button" type="button" onClick={() => {}}>
+              Button item
+            </HeaderMenu.Item>
+            <HeaderMenu.Item button href="#">
+              Item 3
+            </HeaderMenu.Item>
+          </HeaderMenu>,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
 })